Outsource genome search box test
