+++ test/HiGlassComponentTest.js
@@ -33,7 +33,6 @@ import {
   // paperFigure1,
   osmConf,
   geneAnnotationsOnly,
-  geneAnnotationsOnly1,
   annotationsTilesView,
   horizontalAndVerticalMultivec,
   invalidTrackConfig,
@@ -158,54 +157,6 @@ describe('Simple HiGlassComponent', () => {
 
       waitForTransitionsFinished(hgc.instance(), () => {
         done();
-      });
-    });
-  });
-
-  describe('Genome position search box tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(<HiGlassComponent
-        options={{ bounded: false }}
-        viewConfig={geneAnnotationsOnly1}
-      />, { attachTo: div });
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Searches for cdkn2b-as1', (done) => {
-      const firstDomain = hgc.instance().xScales.aa.domain();
-
-      hgc.instance().genomePositionSearchBoxes.aa.onAutocompleteChange({}, 'cdkn2b-as1');
-      hgc.update();
-
-      hgc.instance().genomePositionSearchBoxes.aa.buttonClick();
-      hgc.update();
-
-      waitForJsonComplete(() => {
-        waitForTransitionsFinished(hgc.instance(), () => {
-          const secondDomain = hgc.instance().xScales.aa.domain();
-          // make sure that we zoomed somwhere
-
-          expect(firstDomain[0]).not.toEqual(secondDomain[0]);
-          expect(firstDomain[1]).not.toEqual(secondDomain[1]);
-          done();
-        });
       });
     });
   });
